# app/controllers.go
// Code generated by goagen v1.2.0-dirty, DO NOT EDIT.
//
// API "URL Shortener API": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/rymccue/goa-url-shortener-api/design
// --out=$(GOPATH)/src/github.com/rymccue/goa-url-shortener-api
// --version=v1.2.0-dirty

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// ShortenerController is the controller interface for the Shortener actions.
type ShortenerController interface {
	goa.Muxer
	Analytics(*AnalyticsShortenerContext) error
	Create(*CreateShortenerContext) error
	Delete(*DeleteShortenerContext) error
	Get(*GetShortenerContext) error
	Update(*UpdateShortenerContext) error
}

// MountShortenerController "mounts" a Shortener resource controller on the given service.
func MountShortenerController(service *goa.Service, ctrl ShortenerController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/api/s/:path/analytics", ctrl.MuxHandler("preflight", handleShortenerOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/s", ctrl.MuxHandler("preflight", handleShortenerOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/api/s/:path", ctrl.MuxHandler("preflight", handleShortenerOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewAnalyticsShortenerContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Analytics(rctx)
	}
	h = handleShortenerOrigin(h)
	service.Mux.Handle("GET", "/api/s/:path/analytics", ctrl.MuxHandler("analytics", h, nil))
	service.LogInfo("mount", "ctrl", "Shortener", "action", "Analytics", "route", "GET /api/s/:path/analytics")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateShortenerContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*CreateLinkPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Create(rctx)
	}
	h = handleShortenerOrigin(h)
	service.Mux.Handle("POST", "/api/s", ctrl.MuxHandler("create", h, unmarshalCreateShortenerPayload))
	service.LogInfo("mount", "ctrl", "Shortener", "action", "Create", "route", "POST /api/s")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewDeleteShortenerContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Delete(rctx)
	}
	h = handleShortenerOrigin(h)
	service.Mux.Handle("DELETE", "/api/s/:path", ctrl.MuxHandler("delete", h, nil))
	service.LogInfo("mount", "ctrl", "Shortener", "action", "Delete", "route", "DELETE /api/s/:path")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewGetShortenerContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Get(rctx)
	}
	h = handleShortenerOrigin(h)
	service.Mux.Handle("GET", "/api/s/:path", ctrl.MuxHandler("get", h, nil))
	service.LogInfo("mount", "ctrl", "Shortener", "action", "Get", "route", "GET /api/s/:path")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUpdateShortenerContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*UpdateLinkPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Update(rctx)
	}
	h = handleShortenerOrigin(h)
	service.Mux.Handle("PUT", "/api/s/:path", ctrl.MuxHandler("update", h, unmarshalUpdateShortenerPayload))
	service.LogInfo("mount", "ctrl", "Shortener", "action", "Update", "route", "PUT /api/s/:path")
}

// handleShortenerOrigin applies the CORS response headers corresponding to the origin.
func handleShortenerOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE, PUT, OPTION")
				rw.Header().Set("Access-Control-Allow-Headers", "Content-Type")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// unmarshalCreateShortenerPayload unmarshals the request body into the context request data Payload field.
func unmarshalCreateShortenerPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &createLinkPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}

// unmarshalUpdateShortenerPayload unmarshals the request body into the context request data Payload field.
func unmarshalUpdateShortenerPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &updateLinkPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}